refactor: 일부 함수 이동, 간략화
InputView의 함수 일부를 VendingCalculator로 옮기고 일부는 간략화함
